+++ src/kyber/pke_scheme.rs
@@ -70,7 +70,7 @@ impl<P: PolyParams> KPke<P> {
 
         let mut t_ntt: Vec<PolynomialNTT<P>> = Vec::with_capacity(self.k);
         for i in 0..self.k {
-            let mut pol_temp = PolynomialNTT::<P>::from([0i32; 256]);
+            let mut pol_temp = PolynomialNTT::<P>::from([0i16; 256]);
 
             for (j, poly) in s_ntt.iter().enumerate() {
                 let product = &a_ntt[i * self.k + j] * poly;
@@ -147,7 +147,7 @@ impl<P: PolyParams> KPke<P> {
 
         let mut u = Vec::with_capacity(self.k);
         for (i, poly) in e_1.iter().enumerate() {
-            let mut pol_tmp = PolynomialNTT::<P>::from([0i32; 256]);
+            let mut pol_tmp = PolynomialNTT::<P>::from([0i16; 256]);
             for j in 0..self.k {
                 let product = &a_ntt[j * self.k + i] * &y_ntt[j];
                 pol_tmp += &product;
@@ -156,10 +156,10 @@ impl<P: PolyParams> KPke<P> {
         }
 
         let m_bits = byte_decode(m, 1, P::Q);
-        let mu_coeffs: Vec<i32> = m_bits.into_iter().map(|b| decompress(b, 1, P::Q)).collect();
+        let mu_coeffs: Vec<i16> = m_bits.into_iter().map(|b| decompress(b, 1, P::Q)).collect();
         let mu = Polynomial::<P>::from_slice(&mu_coeffs);
 
-        let mut v_ntt_tmp = PolynomialNTT::<P>::from([0i32; 256]);
+        let mut v_ntt_tmp = PolynomialNTT::<P>::from([0i16; 256]);
         for i in 0..self.k {
             v_ntt_tmp += &(&t_ntt[i] * &y_ntt[i]);
         }
@@ -167,7 +167,7 @@ impl<P: PolyParams> KPke<P> {
 
         let mut c1 = Vec::new();
         for poly in &u {
-            let compressed: Vec<i32> = poly
+            let compressed: Vec<i16> = poly
                 .coeffs
                 .iter()
                 .map(|&c| compress(c, self.d_u, P::Q))
@@ -175,7 +175,7 @@ impl<P: PolyParams> KPke<P> {
             c1.extend(byte_encode(compressed.as_slice(), self.d_u));
         }
 
-        let compressed_v: Vec<i32> = v
+        let compressed_v: Vec<i16> = v
             .coeffs
             .iter()
             .map(|&c| compress(c, self.d_v, P::Q))
@@ -202,7 +202,7 @@ impl<P: PolyParams> KPke<P> {
                 self.d_u,
                 P::Q,
             );
-            let coeffs: Vec<i32> = decode
+            let coeffs: Vec<i16> = decode
                 .into_iter()
                 .map(|val| decompress(val, self.d_u, P::Q))
                 .collect();
@@ -210,7 +210,7 @@ impl<P: PolyParams> KPke<P> {
         }
 
         let decoded_v = byte_decode(c_2, self.d_v, P::Q);
-        let v_coeffs: Vec<i32> = decoded_v
+        let v_coeffs: Vec<i16> = decoded_v
             .into_iter()
             .map(|val| decompress(val, self.d_v, P::Q))
             .collect();
@@ -223,13 +223,13 @@ impl<P: PolyParams> KPke<P> {
             s_ntt.push(PolynomialNTT::<P>::from_slice(coeffs.as_slice()));
         }
 
-        let mut pdt_tmp = PolynomialNTT::<P>::from([0i32; 256]);
+        let mut pdt_tmp = PolynomialNTT::<P>::from([0i16; 256]);
         for i in 0..self.k {
             pdt_tmp += &(&s_ntt[i] * &u_prime[i].to_ntt());
         }
         let w = &v_prime - &Polynomial::<P>::from_ntt(&pdt_tmp);
 
-        let compressed_w: Vec<i32> = w
+        let compressed_w: Vec<i16> = w
             .coeffs
             .iter()
             .map(|&coeff| compress(coeff, 1, P::Q))
